streamlit_app/requirements.txt
streamlit==1.41.1
google-cloud-bigquery==3.27.0
pandas==2.2.3
pyarrow==18.1.0
